Nexus Agents/requirements/requirement.txt
python-dotenv
langchain-groq
langchain-together
langgraph
httpx==0.27.2
langchain-mistralai
pytest
pytest-asyncio
streamlit
browser-use
nest-asyncio